requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiohttp-retry==2.9.1
aiosignal==1.3.2
asgiref==3.8.1
attrs==25.3.0
certifi==2025.1.31
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
frozenlist==1.6.0
h11==0.16.0
idna==3.10
inflection==0.5.1
msgpack==1.1.0
multidict==6.4.3
packaging==25.0
propcache==0.3.1
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
sqlparse==0.5.3
twilio==9.5.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.2
yarl==1.20.0
